src/requirements.txt
attrs==22.1.0
beaker-pyteal==0.3.6
cffi==1.15.1
docstring-parser==0.14.1
iniconfig==1.1.1
msgpack==1.0.4
packaging==21.3
pluggy==1.0.0
py==1.11.0
py-algorand-sdk==1.19.0
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.10.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyteal==0.18.1
python-dotenv==0.21.0
semantic-version==2.10.0
tomli==2.0.1
typing_extensions==4.4.0